fix(server): un handler vacio tapaba un overflow que anulaba el aturdimiento
StunPlayer usaba TickAfter, que hace la resta cruda de dos Longs y desborda
cuando el tick de ahora y el limite caen en mitades opuestas del anillo de
2^32 -- pasa con el uptime de Windows cruzando los ~25 dias. El error 6 se lo
tragaba un handler vacio y la funcion salia devolviendo False: el stun no se
aplicaba, sin efecto visual, sin mensaje y sin log. DeadlinePassed ya existia
para esto desde el plan 06.002; StunPlayer se habia quedado con la version
vieja. Ahora usa la correcta y el handler registra.

Ademas:
- Se completa la cadena de dano a Long. UseArpon declaraba Damage As Integer
  recibiendo GetUserDamageWithItem, que devuelve Long -- el mismo caso de
  NpcDamage. Ensancharlo solo ahi movia el desbordamiento un frame mas abajo
  y encima empeoraba el sintoma: los efectos del golpe ya se habian enviado.
  Tres de las cuatro funciones del motor ya aceptaban Long; el cuello eran los
  wrappers DoDamageToTarget, NpcDoDamageToTarget y RefDoDamageToTarget, mas un
  CInt() explicito al llamar NpcDamageToNpc.
- TieneHechizo deja de tragar el error: devolver False por una excepcion se lee
  como "no tiene el hechizo" y bloquea recetas de crafteo y requisitos de quest.
- La pantalla de GM de intervalos guardaba en intervalo.ini y el server lee
  intervalos.ini: escribia en un archivo que ni existe en disco, asi que los
  ajustes se perdian al reiniciar sin ningun error visible. 25 escrituras.
- clsClan.p_RequiredExp y PoisonedWeaponTimerTick pasan a Long: sin uso hoy,
  pero uno guarda experiencia de clan (la curva llega a 75000) y el otro un
  contador de milisegundos, que desborda a los 33 segundos de encendido.
- Carteles sobre codigo que no se compila: el bloque #If PYMMO = 1 de 394
  lineas que esta en medio del login, y RadarDebugLog. Y queda escrito por que
  el handler de CheckMailString esta vacio a proposito (es la caida natural de
  la funcion, loguear ahi registraria cada validacion exitosa).

Plan ia/plans/2026/agosto/08.005, ola 3.

+++ Codigo/ReferenceUtils.bas
@@ -123,7 +123,7 @@ End Sub
 
 Public Function DoDamageToTarget(ByVal UserIndex As Integer, _
                                  ByRef TargetRef As t_AnyReference, _
-                                 ByVal Damage As Integer, _
+                                 ByVal Damage As Long, _
                                  ByVal DamageType As e_DamageSourceType, _
                                  ByVal ObjIndex As Integer) As e_DamageResult
     If Not IsValidRef(TargetRef) Then
@@ -138,7 +138,7 @@ End Function
 
 Public Function NpcDoDamageToTarget(ByVal NpcIndex As Integer, _
                                     ByRef TargetRef As t_AnyReference, _
-                                    ByVal Damage As Integer, _
+                                     ByVal Damage As Long, _
                                     ByVal DamageType As e_DamageSourceType, _
                                     ByVal ObjIndex As Integer) As e_DamageResult
     If Not IsValidRef(TargetRef) Then
@@ -153,7 +153,7 @@ End Function
 
 Public Function RefDoDamageToTarget(ByRef SourceRef As t_AnyReference, _
                                     ByRef TargetRef As t_AnyReference, _
-                                    ByVal Damage As Integer, _
+                                     ByVal Damage As Long, _
                                     ByVal DamageType As e_DamageSourceType, _
                                     ByVal ObjIndex As Integer) As e_DamageResult
     If Not IsValidRef(SourceRef) Then
